Move index migration to version18
update changelog and version
