+++ src/Hangfire.Mongo.Sample/packages.config
@@ -3,11 +3,18 @@
   <package id="Antlr" version="3.5.0.2" targetFramework="net45" />
   <package id="SharpCompress" version="0.25.1" targetFramework="net46" />
   <package id="System.Memory" version="4.5.4" targetFramework="net46" />
+  <package id="MongoDB.Driver" version="2.11.0" targetFramework="net46" />
+  <package id="MongoDB.Driver.Core" version="2.11.0" targetFramework="net46" />
+  <package id="System.Net.Http" version="4.3.4" targetFramework="net46" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net46" />
   <package id="bootstrap" version="4.4.1" targetFramework="net46" />
   <package id="Crc32C.NET" version="1.0.5.0" targetFramework="net46" />
   <package id="DnsClient" version="1.3.2" targetFramework="net46" />
-  <package id="Hangfire.Core" version="1.7.11" targetFramework="net46" />
+  <package id="Hangfire.Core" version="1.7.12" targetFramework="net46" />
   <package id="jQuery" version="3.4.1" targetFramework="net46" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.3" targetFramework="net46" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.7" targetFramework="net46" />
@@ -17,9 +24,7 @@
   <package id="Microsoft.Owin" version="4.1.0" targetFramework="net46" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="4.1.0" targetFramework="net46" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net45" />
-  <package id="MongoDB.Bson" version="2.10.4" targetFramework="net46" />
-  <package id="MongoDB.Driver" version="2.10.4" targetFramework="net46" />
-  <package id="MongoDB.Driver.Core" version="2.10.4" targetFramework="net46" />
+  <package id="MongoDB.Bson" version="2.11.0" targetFramework="net46" />
   <package id="MongoDB.Libmongocrypt" version="1.0.0" targetFramework="net46" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net46" />
   <package id="Owin" version="1.0" targetFramework="net45" />
