+++ src/Hangfire.Mongo.Tests/Migration/Mongo/MigrationFacts.cs
@@ -37,6 +37,7 @@ namespace Hangfire.Mongo.Tests.Migration.Mongo
         [InlineData("Hangfire-Mongo-Schema-015.zip", true)]
         [InlineData("Hangfire-Mongo-Schema-016.zip", true)]
         [InlineData("Hangfire-Mongo-Schema-017.zip", true)]
+        [InlineData("Hangfire-Mongo-Schema-018.zip", true)]
         public void Migrate_Full_Success(string seedFile, bool assertCollectionHasItems)
         {
             var dbContext =
