+++ src/Hangfire.Mongo/Hangfire.Mongo.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <VersionPrefix>0.7.11</VersionPrefix>
+    <VersionPrefix>0.7.12</VersionPrefix>
     <TargetFrameworks>net46;net452;netstandard2.0</TargetFrameworks>
     <NoWarn>$(NoWarn);CS0618</NoWarn>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
@@ -17,19 +17,15 @@
     <GenerateAssemblyFileVersionAttribute>false</GenerateAssemblyFileVersionAttribute>
     <GenerateAssemblyInformationalVersionAttribute>false</GenerateAssemblyInformationalVersionAttribute>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Copyright>Copyright 2019 © Sergey Zwezdin</Copyright>
+    <Copyright>Copyright 2020 © Sergey Zwezdin</Copyright>
     <title>Hangfire MongoDB Storage</title>
     <owners>Sergey Zwezdin</owners>
     <Description>MongoDB storage implementation for Hangfire (background job system for ASP.NET applications).</Description>
     <PackageTags>Hangfire AspNet OWIN MongoDB Long-Running Background Fire-And-Forget Delayed Recurring Tasks Jobs Scheduler Threading Queues</PackageTags>
-    <PackageReleaseNotes>0.7.11
-      - Align version number with Hangfire.Core
-      - Update dependencies
-         - target netstandard2.0
-         - "MongoDB.Driver" Version="2.10.4"
-         - "Hangfire.Core" Version="1.7.11"
-      - Make extendable
-         - Everything can now be hooked into and extended
+    <PackageReleaseNotes>0.7.12
+      - Update to latest Hangfire.Core (v1.7.12)
+      - Update to latest Mongo drivers (v2.11.0)
+      - Add index (PR #247)
 </PackageReleaseNotes>
     <!--<PackageLicenseUrl>https://raw.githubusercontent.com/sergun/Hangfire.Mongo/master/LICENSE</PackageLicenseUrl>-->
     <PackageProjectUrl>https://github.com/sergeyzwezdin/Hangfire.Mongo</PackageProjectUrl>
@@ -39,8 +35,8 @@
     <None Include="app.config" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Hangfire.Core" Version="1.7.11" />
-    <PackageReference Include="MongoDB.Driver" Version="2.10.4" />
+    <PackageReference Include="Hangfire.Core" Version="1.7.12" />
+    <PackageReference Include="MongoDB.Driver" Version="2.11.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
 </Project>
