+++ src/Hangfire.Mongo/Properties/AssemblyInfo.cs
@@ -29,8 +29,8 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.7.11")]
-[assembly: AssemblyFileVersion("0.7.11")]
-[assembly: AssemblyInformationalVersion("0.7.11")]
+[assembly: AssemblyVersion("0.7.12")]
+[assembly: AssemblyFileVersion("0.7.12")]
+[assembly: AssemblyInformationalVersion("0.7.12")]
 [assembly: InternalsVisibleTo("Hangfire.Mongo.Tests")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
